Move runtime spec into submodule
This allows to add the image spec as parallel module later on.
