+++ src/runtime/hooks.rs
