+++ src/runtime/vm.rs
